[uss_qualifier/scenarios/netrid] Move recent position timings checks to display_data_evaluator from common_data_dictionary
